Revamp code structure and fix typo errors.
